Adição do dockerfile e criação dos testes de unidade e de integração.

+++ src/main/java/com/challenges/desafiopetrobrasbackend/repository/EventosRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventosRepository extends JpaRepository<Eventos, Integer> {
+public interface EventosRepository extends JpaRepository<Eventos, Long> {
     Page<Eventos> findAllByDeleted(String deleted, Pageable pageable);
 }
